requirements.txt
pandas
numpy
torch
matplotlib
seaborn
scikit-learn
dill
fastapi
uvicorn
gunicorn
pydantic
python-box
pyYAML
-e .